upload/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.3.0
beanie==1.29.0
certifi==2025.4.26
cffi==1.17.1
click==8.2.1
cryptography==45.0.2
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.115.12
h11==0.16.0
idna==3.10
lazy-model==0.2.0
minio==7.2.15
motor==3.7.1
passlib==1.7.4
pathvalidate==3.3.1
pyasn1==0.4.8
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.13.0
python-dotenv==1.1.0
python-jose==3.4.0
redis==6.1.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
toml==0.10.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
